Split py_machine into header and source files; Fixed compilation errors in FFNN and DensityMatrices
